+++ ostap/fitting/tests/test_fitting_models.py
@@ -273,7 +273,7 @@ def test_crystalball_RS () :
 def test_crystalball_DS () :
     
     logger = getLogger ( 'test_crystalball_DS' )
-    logger.info ('Test CrystalBallDS_pdf: double-sided Crystal Ball function' )
+    logger.info ('Test CB2_pdf: double-sided Crystal Ball function' )
     model = Models.Fit1D (
         signal = Models.CB2_pdf ( name   = 'CB2'              , 
                                   xvar   = mass               ,
@@ -379,7 +379,7 @@ def test_crystalballa () :
 def test_crystalball_DSA () :
     
     logger = getLogger ( 'test_crystalball_DSA' )
-    logger.info ('Test CrystalBallDSA_pdf: double-sided Crystal Ball function' )
+    logger.info ('Test CB2A_pdf: double-sided Crystal Ball function' )
     model = Models.Fit1D (
         signal = Models.CB2A_pdf ( name   = 'CB2A'               , 
                                    xvar   = mass                 ,
@@ -410,6 +410,43 @@ def test_crystalball_DSA () :
 
     models.add ( model )
     results.append ( result  )
+    
+# =============================================================================
+## double sided CrystalBall with asymmetric core 
+# =============================================================================
+def test_crystalball_DSE () :
+    
+    logger = getLogger ( 'test_crystalball_DSS' )
+    logger.info ('Test CB2E_pdf: double-sided Crystal Ball function' )
+    model = Models.Fit1D (
+        signal = Models.CB2E_pdf ( name   = 'CB2E'               , 
+                                   xvar   = mass                 ,
+                                   nL     = 10                   , 
+                                   alphaL = (1.5,0.5,3)          , 
+                                   alphaR = (1.4,0.5,3)          , 
+                                   sigma  = signal_gauss.sigma   ,  
+                                   mean   = signal_gauss.mean    ,
+                                   psi    = ( 0.03 , -0.5 , +0.5 ) ) , 
+        background = background   ,
+        S = S , B = B 
+    )
+    
+    model.S = NS 
+    model.B = NB
+    
+    with rooSilent() : 
+        result, frame = model. fitTo ( dataset0 )
+        model.signal.aL.release()
+        model.signal.aR.release()
+        result, frame = model. fitTo ( dataset0 )
+        model.signal.aL.fix(1.5) 
+        model.signal.aR.fix(1.5)    
+        result, frame = model. fitTo ( dataset0 )
+        
+    make_print ( model, result , 'Double-sided Crystal Ball/E model' , logger )
+
+    models.add ( model )
+    results.append ( result  )    
 
 # ==========================================================================
 ## Apollonios
@@ -2202,6 +2239,12 @@ if '__main__' == __name__ :
     with timing ('test_crystalball_DSA' , logger ) :
         test_crystalball_DSA () 
 
+    ## double side Crystal Ball/E                  + background
+    with timing ('test_crystalball_DSE' , logger ) :
+        test_crystalball_DSE () 
+        
+    """ 
+        
     ## Apollonios function                       + background 
     with timing ('test_apollonios'     , logger ) :
         test_apollonios     () 
@@ -2378,7 +2421,8 @@ if '__main__' == __name__ :
     ## Hypatia                                     + background 
     with timing ('test_hypatia'           , logger ) :
         test_hypatia           ()
-
+    
+    """
     ## check finally that everything is serializeable:
     with timing ('test_db'             , logger ) :
         test_db ()